ML-22 Training network

+++ src/main/app/business/audio/generation/preprocessing/audio_retrieval.py
@@ -3,17 +3,17 @@ from numpy import ndarray
 from werkzeug.datastructures import FileStorage
 
 
-def retrieve_via_path(path: str, required_sample_rate: int) -> tuple[ndarray, int]:
+def retrieve_via_path(path: str, required_sample_rate: float) -> tuple[ndarray, float]:
     audio_data, sr = librosa.load(path)
     return _normalize(audio_data, sr, required_sample_rate), required_sample_rate
 
 
-def retrieve_via_storage(storage: FileStorage, required_sample_rate: int) -> tuple[ndarray, int]:
+def retrieve_via_storage(storage: FileStorage, required_sample_rate: float) -> tuple[ndarray, float]:
     audio_data, sr = librosa.load(storage)
     return _normalize(audio_data, sr, required_sample_rate), sr
 
 
-def _normalize(audio_data, sr: int, required_sample_rate: int) -> ndarray:
+def _normalize(audio_data, sr: int, required_sample_rate: float) -> ndarray:
     # Normalize the audio
     audio_data /= max(abs(audio_data))
     # Trim leading and trailing silence
